Adopt acceptance tests as go tests
- The godog CLI is now depricated.

- To make the IDE work, you need to add the tag to the workspace configuration.
