+++ Makefile
@@ -76,7 +76,7 @@ acceptance-test: build ## run acceptance tests
 	rm -rf ./test/tmp
 	go build -cover -o layli ./cmd/layli
 	mkdir -p ./test/tmp/coverage
-	cd test && GOCOVERDIR=tmp/coverage godog run
+	cd test && GOCOVERDIR=tmp/coverage go test -tags acceptance
 	
 .PHONY: coverage-report
 coverage-report: ## collate the coverage data
